refactor: 컨택 페이지 폼 로직 분리(useContactForm) 및 푸터 링크 업데이트
