+++ src/components/layout/Footer.tsx
@@ -25,14 +25,17 @@ const MAIN_LINKS: MainLink[] = [
 const DOCUMENT_LINKS = [
   {
     icon: "📄",
-    label: "와이어 프레임/프로토타입",
+    label: "와이어 프레임",
     href: "https://www.figma.com/design/JG4gfkkbaCxxnanJaEfLti/2026-%EC%9D%B4%EB%A0%A5%EC%84%9C--%ED%8F%AC%ED%8A%B8%ED%8F%B4%EB%A6%AC%EC%98%A4?node-id=0-1&t=UxVLKsCl0Rhyn89e-1",
     external: true,
   },
-  { icon: "📄", label: "기능 명세서", href: "#", external: true },
   { icon: "📄", label: "백로그", href: "/backlog", external: false },
-  { icon: "📄", label: "기술 구현 문서", href: "#", external: true },
-  { icon: "📄", label: "트러블 슈팅", href: "#", external: true },
+  {
+    icon: "📄",
+    label: "개인 대시보드",
+    href: "/admin/profile-settings",
+    external: true,
+  },
 ] as const;
 
 function FooterLink({
